refactored whole repo to have proper structure. added scalastyle and plugins and xml file etc..
