+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.investmenttaxreliefsubscription.controllers.MicroserviceHelloWorld.hello
+GET        /hello-world            controllers.TestController.hello
