+++ conf/application.conf
@@ -1,3 +1,16 @@
+# Copyright 2016 HM Revenue & Customs
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 # This is the main configuration file for the application.
 # ~~~~~
@@ -15,7 +28,7 @@ appName=investment-tax-relief-subscription
 # ~~~~~
 # The secret key is used to secure cryptographics functions.
 # If you deploy your application to several instances be sure to use the same key!
-application.secret="2uQAMDi5VNrQhwTmKtIZjFrsrQ22SsBWJ526U3s5gHG6uWQNClic25eJmcmPWhIa"
+application.secret="bDuxCZk9I7jlOhYCC8DDixh9i2ds37Sncqp6cP0tyuiEYYhNRPh0MZOrVGU740TP"
 
 # Session configuration
 # ~~~~~
@@ -31,7 +44,7 @@ application.langs="en"
 # ~~~~~
 # Define the Global object class for this application.
 # Default to Global in the root package.
-application.global=uk.gov.hmrc.investmenttaxreliefsubscription.MicroserviceGlobal
+application.global=config.MicroserviceGlobal
 
 # Router
 # ~~~~~
@@ -62,7 +75,7 @@ controllers {
       needsAuditing = false
     }
 
-  uk.gov.hmrc.investmenttaxreliefsubscription.controllers.MicroserviceHelloWorld = {
+  controllers.TestController = {
       needsAuth = false
       needsLogging = false
       needsAuditing = false
@@ -138,5 +151,3 @@ microservice {
 
     }
 }
-
-
