+++ project/plugins.sbt
@@ -9,3 +9,9 @@ addSbtPlugin("uk.gov.hmrc" % "sbt-git-versioning" % "0.8.0")
 addSbtPlugin("uk.gov.hmrc" % "sbt-distributables" % "0.11.0")
 
 addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.3.10")
+
+resolvers += Resolver.url("scoverage-bintray", url("https://dl.bintray.com/sksamuel/sbt-plugins/"))(Resolver.ivyStylePatterns)
+
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.3.5")
+
+addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "0.8.0")
